Three-way ACS Tables for New York

Table of counts by CD from the 2018 one-year ACS in New York. We load the variables that roughly partition gender, age, and race. We get another three way table of gender, age, and education. That is, this is the output of get_acs_cces(acscodes_age_sex_race) and get_acs_cces(acscodes_age_sex_educ).

acs_educ_NY

acs_race_NY

Format

An object of class tbl_df (inherits from tbl, data.frame) with 1890 rows and 6 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 4320 rows and 6 columns.

Source

Kyle Walker and Matt Herman (2021). tidycensus: Load US Census Boundary and Attribute Data as 'tidyverse' and 'sf'-Ready Data Frames. R package version 0.11.4. https://CRAN.R-project.org/package=tidycensus

Examples

 library(tibble)
 acs_race_NY
#> # A tibble: 4,320 × 6
#>     year cd    female race  age               count
#>    <dbl> <chr>  <int> <fct> <fct>             <dbl>
#>  1  2018 NY-01      0 Black 18 to 24 years      826
#>  2  2018 NY-01      0 Black 18 to 24 years     1828
#>  3  2018 NY-01      0 Black 25 to 34 years     1127
#>  4  2018 NY-01      0 Black 25 to 34 years     1298
#>  5  2018 NY-01      0 Black 35 to 44 years     2696
#>  6  2018 NY-01      0 Black 45 to 64 years     3779
#>  7  2018 NY-01      0 Black 45 to 64 years     2153
#>  8  2018 NY-01      0 Black 65 years and over   781
#>  9  2018 NY-01      0 Black 65 years and over   624
#> 10  2018 NY-01      0 Black 65 years and over   133
#> # … with 4,310 more rows
 acs_educ_NY
#> # A tibble: 1,890 × 6
#>     year cd    female educ         age            count
#>    <dbl> <chr>  <int> <fct>        <fct>          <dbl>
#>  1  2018 NY-01      0 HS or Less   18 to 24 years   394
#>  2  2018 NY-01      0 HS or Less   18 to 24 years  2066
#>  3  2018 NY-01      0 HS or Less   18 to 24 years 10330
#>  4  2018 NY-01      0 Some College 18 to 24 years 13334
#>  5  2018 NY-01      0 Some College 18 to 24 years  2064
#>  6  2018 NY-01      0 4-Year       18 to 24 years  4584
#>  7  2018 NY-01      0 Post-Grad    18 to 24 years   155
#>  8  2018 NY-01      0 HS or Less   25 to 34 years  1718
#>  9  2018 NY-01      0 HS or Less   25 to 34 years  1349
#> 10  2018 NY-01      0 HS or Less   25 to 34 years 10853
#> # … with 1,880 more rows