ACS codes for partitions

Each vector defines a partition, or at least an attempt at one. Each element of the vector is a code for a ACS variable, which can be used, for example in the variables argument of tidycensus::get_acs(). Some of these will be collapsed for MRP to form a common denominator with the CCES; see keyvalue

acscodes_age_sex_race

acscodes_age_sex_educ

acscodes_age_sex_educ

Format

acscodes_age_sex_race

There are 160 codes in acscodes_age_sex_rac because they specify cells interacting age (10 bins), sex (2 bins), and race/ethnicity (8 bins).

acscoes_age_sex_educ

There are 70 codes in acscodes_age_sex_edu because they specify cells interacting age (5 bins), sex (2 bins), and education (7 bins).

acscodes_sex_educ_race

There are 48 codes in acscodes_sex_educ_race because they specify cells interacting sex (2 bins), education (3 bins), and race (8 bins).

See also

Examples

head(acscodes_age_sex_race)

#> [1] "B01001B_007" "B01001B_008" "B01001B_009" "B01001B_010" "B01001B_011"
#> [6] "B01001B_012"

head(acscodes_age_sex_educ)

#> [1] "B15001_004" "B15001_005" "B15001_006" "B15001_007" "B15001_008"
#> [6] "B15001_009"

head(acscodes_sex_educ_race)

#> [1] "C15002B_004" "C15002B_005" "C15002B_006" "C15002B_009" "C15002B_010"
#> [6] "C15002B_011"