Collapse CCES data to be analyzed in binomial model

Currently only is compatible with question of type "yesno".

build_counts(data, model_ff, multiple_qIDs = FALSE)

Arguments

data
A cleaned CCES dataset, e.g. from ccc_std_demographics which is then combined with outcome and contextual data in cces_join_slim. Currently, it assumes the outcome is named response. This variable need not be numeric or binarized, but it will be binarized with yesno_to_binary.

model_ff
the model formula used to fit the multilevel regression model. Currently only expects an binomial, of the brms form y|trials(n) ~ x1 + x2 + (1|x3). Only the RHS will be used but the LHS is necessary.

multiple_qIDs
Does the data contain multiple outcomes in long form and therefore require the counts to be built for each outcome? Defaults to FALSE.

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

ccc_samp_std <- ccc_samp %>%
  mutate(response = sample(c("For", "Against"), size = n(), replace = TRUE)) %>%
  ccc_std_demographics()

ff <- "yes | trials(n_response) ~ age + gender + educ + (1|cd)"
ccc_samp_out <- build_counts(ccc_samp_std,
                             model_ff = ff)

ccc_samp_out

#> # A tibble: 996 x 6
#>      age gender educ       cd    n_response   yes
#>    <dbl> <fct>  <fct>      <chr>      <int> <int>
#>  1    18 Male   HS or Less FL-25          1     1
#>  2    18 Male   HS or Less NC-09          1     0
#>  3    18 Male   HS or Less PA-11          1     0
#>  4    18 Female HS or Less AL-05          1     1
#>  5    18 Female HS or Less CT-05          1     1
#>  6    18 Female HS or Less IN-05          1     0
#>  7    18 Female HS or Less MD-05          1     1
#>  8    18 Female HS or Less MN-04          1     1
#>  9    18 Female HS or Less OR-02          1     1
#> 10    18 Female HS or Less PA-05          1     0
#> # … with 986 more rows