Sample 2018 Common Content

1000 rows from th 2018 CCES Common Content. Contains all columns, retaining the haven_labelled class. This is doi:10.7910/DVN/ZSBZ7K/H5IDTA originally read in using read_dta via get_cces_dv(). To search for variable by its content, questionr::lookfor (or rcces::vartab) is a useful option (see example).

cc18_samp

Format

An object of class tbl_df (inherits from tbl, data.frame) with 1000 rows and 526 columns.

Source

"Brian Schaffner; Stephen Ansolabehere; Sam Luks, 2019, "CCES Common Content, 2018", doi:https://doi.org/10.7910, Harvard Dataverse V6.

Details

See the 2018 Codebook in the DOI below for question wording of each column. To use the harmonized and easier-to-use versions of common variables like geography, partisan ID, demographics and vote choice, use the cumulative common content, of which 2018 is a subset. A sample of the cumulative is contained in this package as ?ccc_samp.

Examples

# use questionr::lookfor to search the label and labels
questionr::lookfor(cc18_samp, "Trump")

#>                variable                                 label
#> 91            CC18_308a       Job approval -- President Trump
#> 149           CC18_334C Ideological Placement -- Donald Trump
#> 162            CC18_335                Trump Russia collusion
#> 430 CC18_app_dtrmp_post          President Trump Job Approval

# all data
cc18_samp

#> # A tibble: 1,000 x 526
#>     year case_id commonweight commonpostweight vvweight vvweight_post tookpost
#>    <dbl>   <dbl>        <dbl>            <dbl>    <dbl>         <dbl> <dbl+lb>
#>  1  2018  4.15e8        1.47             1.48    NA            NA      2 [Yes]
#>  2  2018  4.14e8        1.33             1.22     1.51          1.29   2 [Yes]
#>  3  2018  4.12e8        0.641            0.559    0.602         0.594  2 [Yes]
#>  4  2018  4.14e8        1.35             1.19     1.18          1.17   2 [Yes]
#>  5  2018  4.12e8        0.870            0.762    0.842         0.742  2 [Yes]
#>  6  2018  4.12e8        0.949            0.789    1.07          0.983  2 [Yes]
#>  7  2018  4.17e8        0.688            0.585    0.717         0.716  2 [Yes]
#>  8  2018  4.15e8        0.814            0.616   NA            NA      2 [Yes]
#>  9  2018  4.17e8        1.12             0.926    0.991         0.929  2 [Yes]
#> 10  2018  4.13e8        1.05             0.877    0.346         0.372  2 [Yes]
#> # … with 990 more rows, and 519 more variables: CCEStake <dbl+lbl>,
#> #   birthyr <dbl>, gender <dbl+lbl>, educ <dbl+lbl>, race <dbl+lbl>,
#> #   race_other <chr>, hispanic <dbl+lbl>, marstat <dbl+lbl>,
#> #   multrace_1 <dbl+lbl>, multrace_2 <dbl+lbl>, multrace_3 <dbl+lbl>,
#> #   multrace_4 <dbl+lbl>, multrace_5 <dbl+lbl>, multrace_8 <dbl+lbl>,
#> #   multrace_97 <dbl+lbl>, multrace_98 <dbl+lbl>, multrace_99 <dbl+lbl>,
#> #   CC18_354a_1 <dbl+lbl>, CC18_354a_2 <dbl+lbl>, CC18_354a_3 <dbl+lbl>,
#> #   CC18_354a_4 <dbl+lbl>, CC18_354a_5 <dbl+lbl>, CC18_354a_6 <dbl+lbl>,
#> #   CC18_354a_7 <dbl+lbl>, CC18_354a_8 <dbl+lbl>, CC18_354a_9 <dbl+lbl>,
#> #   CC18_354a_10 <dbl+lbl>, CC18_354a_11 <dbl+lbl>, CC18_354a_12 <dbl+lbl>,
#> #   CC18_354a_t <chr>, CC18_354b_1 <dbl+lbl>, CC18_354b_2 <dbl+lbl>,
#> #   CC18_354b_3 <dbl+lbl>, CC18_354b_4 <dbl+lbl>, CC18_354b_5 <dbl+lbl>,
#> #   CC18_354b_6 <dbl+lbl>, CC18_354b_7 <dbl+lbl>, CC18_354b_8 <dbl+lbl>,
#> #   CC18_354b_9 <dbl+lbl>, CC18_354b_10 <dbl+lbl>, CC18_354b_11 <dbl+lbl>,
#> #   CC18_354b_12 <dbl+lbl>, CC18_354b_13 <dbl+lbl>, CC18_354b_14 <dbl+lbl>,
#> #   CC18_354b_15 <dbl+lbl>, CC17_3534_t <chr>, comptype <dbl+lbl>,
#> #   add_confirm <dbl+lbl>, inputzip <dbl+lbl>, votereg <dbl+lbl>,
#> #   votereg_f <dbl+lbl>, regzip <dbl+lbl>, OScode <dbl+lbl>,
#> #   CompRating <dbl+lbl>, region <dbl+lbl>, sample_group <dbl+lbl>,
#> #   ccesmodule <dbl+lbl>, CC18_300_1 <dbl+lbl>, CC18_300_2 <dbl+lbl>,
#> #   CC18_300_3 <dbl+lbl>, CC18_300_4 <dbl+lbl>, CC18_300_5 <dbl+lbl>,
#> #   CC18_300_6 <dbl+lbl>, CC18_300b <dbl+lbl>, CC18_300c <dbl+lbl>,
#> #   CC18_300d_1 <dbl+lbl>, CC18_300d_2 <dbl+lbl>, CC18_300d_3 <dbl+lbl>,
#> #   CC18_300d_4 <dbl+lbl>, CC18_300d_5 <dbl+lbl>, CC18_301 <dbl+lbl>,
#> #   CC18_302 <dbl+lbl>, CC18_303_1 <dbl+lbl>, CC18_303_2 <dbl+lbl>,
#> #   CC18_303_3 <dbl+lbl>, CC18_303_4 <dbl+lbl>, CC18_303_5 <dbl+lbl>,
#> #   CC18_303_6 <dbl+lbl>, CC18_303_7 <dbl+lbl>, CC18_303_8 <dbl+lbl>,
#> #   CC18_303_9 <dbl+lbl>, CC18_303_10 <dbl+lbl>, CC18_303_11 <dbl+lbl>,
#> #   CC18_308a <dbl+lbl>, CC18_308b <dbl+lbl>, CC18_308c <dbl+lbl>,
#> #   CC18_308d <dbl+lbl>, CC18_308e <dbl+lbl>, CC18_309a <dbl+lbl>,
#> #   CC18_309b <dbl+lbl>, CC18_309c <dbl+lbl>, CC18_309d <dbl+lbl>,
#> #   CC18_310a <dbl+lbl>, CC18_310b <dbl+lbl>, CC18_310c <dbl+lbl>,
#> #   CC18_310d <dbl+lbl>, CC18_311a <dbl+lbl>, CC18_311b <dbl+lbl>,
#> #   CC18_311c <dbl+lbl>, CC18_316 <dbl+lbl>, …