Sample cumulative CCES data

The cumulative CCES stacks CCES common content for all years and harmonizes the variables, which makes it ideal for using it for MRP. This is a sample for illustration; see get_cces_dv to get the full data.

ccc_samp

Format

A data frame with 1000 observations. See the CCES cumulative codebook for more explanation of the variables

year

Year of the common content

case_id

Respondent identifier (unique within each year)

state

State (in the form of state.name)

st

State Abbreviation (in the form of state.abb)

cd

Congressional district at the time of the survey

zipcode

Zipcode (See codebook)

county_fips

County FIPS code (See codebook)

age

Age, in integers

race

Race and ethnicity

hispanic

Hispanic identifier

educ

Education level

faminc

Family Income

marstat

Martial status

newsint

Frequency of following the news

vv_turnout_gvm

Validated turnout in general election

voted_pres_16

Self-reported vote choice for 2016

economy_retro

Opinion on retrospective economy

Source

Kuriwaki, Shiro, 2018, "Cumulative CCES Common Content (2006-2018)", https://doi.org/10.7910/DVN/II2DB6, Harvard Dataverse

Details

This is encoded as a RDS file with some variables stored in the Stata-based integer + labelled class instead of as factors. See the CCES cumulative guide for the difference between the two and how to go from one to another.

Examples

library(ccesMRPprep)
ccc_samp

#> # A tibble: 1,000 x 17
#>     year case_id state st    cd    zipcode county_fips   age    race hispanic
#>    <int>   <int> <chr> <chr> <chr> <chr>   <chr>       <int> <int+l> <int+lb>
#>  1  2006  439254 Neva… NV    NV-2  89703   32510          20 1 [Whi…       NA
#>  2  2006  440689 Cali… CA    CA-2  96067   06093          71 1 [Whi…       NA
#>  3  2006  445845 Colo… CO    CO-5  80922   08041          39 1 [Whi…       NA
#>  4  2006  452572 Minn… MN    MN-5  55428   27053          74 1 [Whi…       NA
#>  5  2006  498451 Iowa  IA    IA-2  50060   19185          53 1 [Whi…       NA
#>  6  2006  502543 Tenn… TN    TN-5  37206   47037          50 1 [Whi…       NA
#>  7  2006  523050 West… WV    WV-2  25312   54039          55 1 [Whi…       NA
#>  8  2006  523861 Illi… IL    IL-3  60629   17031          51 1 [Whi…       NA
#>  9  2006  532881 Minn… MN    MN-7  55396   27143          44 1 [Whi…       NA
#> 10  2006  553375 Ohio  OH    OH-11 44112   39035          53 1 [Whi…       NA
#> # … with 990 more rows, and 7 more variables: educ <int+lbl>, faminc <fct>,
#> #   marstat <int+lbl>, newsint <int+lbl>, vv_turnout_gvm <fct>,
#> #   voted_pres_16 <fct>, economy_retro <int+lbl>