Identifiers for datasets in CCES Dataverse

Identifiers for datasets in CCES Dataverse

cces_dv_ids

Format

A tibble with one row per dataverse dataset. The doi, filename, and server can uniquely define a file in ?dataverse::get_file().

cces_name

The unique name/shorthand for this dataset

year

Specific year the data surveys, if there is one unique year

doi

The DOI for the dataverse dataset (instead of the specific file). The URL is therefore https://doi.org/ followed by the DOI

filename

The name of the data file, as it appears in the dataverse.

server

The particular dataverse instance.

caseid_var

The variable in that dataset that specifies the unique case (respondent/person) identifier. Use a string.

Examples

 cces_dv_ids
#> # A tibble: 19 × 6
#>    cces_name   year doi                filename                caseid_var server
#>    <chr>      <dbl> <chr>              <chr>                   <chr>      <chr> 
#>  1 cumulative    NA 10.7910/DVN/II2DB6 cumulative_2006-2023.d… case_id    datav…
#>  2 2006        2006 10.7910/DVN/Q8HC9N cces_2006_common.tab    v1000      datav…
#>  3 2007        2007 10.7910/DVN/OOXTJ5 CCES07_OUTPUT.sav       caseid     datav…
#>  4 2008        2008 10.7910/DVN/YUYIVB cces_2008_common.tab    V100       datav…
#>  5 2009        2009 10.7910/DVN/KKM9UK cces09_cmn_output_2.dta v100       datav…
#>  6 2010        2010 10.7910/DVN/VKKRWA cces_2010_common_valid… V100       datav…
#>  7 2011        2011 10.7910/DVN/GLRXZ0 CCES11_Common_OUTPUT.t… V100       datav…
#>  8 2012        2012 10.7910/DVN/HQEVPK CCES12_Common_VV.tab    V101       datav…
#>  9 2013        2013 10.7910/DVN/KPP85M Common Content Data.tab caseid     datav…
#> 10 2014        2014 10.7910/DVN/XFXJVY CCES14_Common_Content_… V101       datav…
#> 11 2015        2015 10.7910/DVN/SWMWX8 CCES15_Common_OUTPUT_J… V101       datav…
#> 12 2016        2016 10.7910/DVN/GDF6Z0 CCES16_Common_OUTPUT_F… V101       datav…
#> 13 2017        2017 10.7910/DVN/3STEZY Common Content Data.tab V101       datav…
#> 14 2018        2018 10.7910/DVN/ZSBZ7K cces18_common_vv.dta    caseid     datav…
#> 15 2019        2019 10.7910/DVN/WOT7O8 CCES19_Common_OUTPUT.t… caseid     datav…
#> 16 2020        2020 10.7910/DVN/E9N6PH CES20_Common_OUTPUT_vv… caseid     datav…
#> 17 2021        2021 10.7910/DVN/OPQOCU CES21_Common_OUTPUT.tab caseid     datav…
#> 18 2022        2022 10.7910/DVN/PR4L8P CCES22_Common_OUTPUT_v… caseid     datav…
#> 19 2023        2023 10.7910/DVN/JQJTCC CCES23_Common_OUTPUT.t… caseid     datav…