Congressional District level information by The Downballot

Some of the most consequential variables to include in MRP are measured at the district-level. We include one such data for congressional districts. The data is collected by The Downballot and Kiernan Park-Egan.

cd_info_2008

cd_info_2012

cd_info_2016

cd_info_2018

cd_info_2020

cd_info_2022

cd_info_2024

elec_NY

Format

Each cd_info_20** is a dataframe with the 435 Congressional Districts, one row per cd.

year

The year for the district map. A congressional district's actual geography can change year to year. Lines represent the contemporaneous district geography, so that cd_info_2016 uses 2016 maps and cd_info_2020 uses 2020 maps. Corresponds to line in cd_info_long. This work relies on the hard work of assembling precinct results by Daily Kos.

cd

District code. The formatting corresponds to the CCES cumulative coding of cd: a two-letter abbreviation for the state followed by a dash, and the district number padded with zeros to the left to be of length 2. At-large districts like Delaware are given a "-01" for the district number. See to_cd()

presvotes_total

In presidential years, the total number of votes cast for the office of President that year.

presvotes_DR

Same as presvotes_total but only the sum of Democratic and Republican candidate's votes

pct_trump, pct_romney, pct_mccain

The two-party voteshare of Republican presidential candidates in that district for the given year. E.g. the pct_mccain data for cd_info_2018 represents the percent of the vote by McCain in 2008 for that district under 2018 lines.
pct_trump denotes the 2016 election for cd_info_2018 and cd_info_2016.
pct_trump denotes the 2020 election for cd_info_2020, cd_info_2022, and cd_info_2024.
pct_trump16 denotes the 2016 result for cd_info_2020.

dailykos_name

The unique descriptive name for the district code in 2018 given by Daily Kos (later renamed to The Downballot). Some edits are made for changing district. See Source for full citation.

largest_place

The largest place in the district code in 2018 given by Daily Kos. Multiple districts may have the same largest place.

Source

The Downballot (formerly Daily Kos Elections), https://www.the-downballot.com/p/data

The Daily Kos Elections naming guide to the nation's congressional districts. https://bit.ly/2XsFI5W

Daily Kos, "2008 results for districts used in 2006, 2008, 2010" https://bit.ly/4entUrV

Kiernan Park-Egan, "U.S. Presidential Election Results by Congressional District, 1952 to 2020" https://bit.ly/4fk6UKx (used for 2008 values only when Daily Kos has missing data)

Daily Kos, "2008, 2012 results for districts used in 2012, 2014" https://bit.ly/3N4PDZK

Daily Kos, "2008, 2012, & 2016 results for districts used in 2018." https://bit.ly/3bXtAPB

Daily Kos, "2012, 2016 & 2020 presidential election results for congressional districts in 2020", https://bit.ly/3DRhPcj

Daily Kos, 2020 presidential election results by later congressional districts:

2022 congressional districts: https://bit.ly/4gLYnBK

2024 congressional districts: https://bit.ly/47KTvZw

Daily Kos, congressional district geography and most populous places:

119th Congress: https://bit.ly/geography_119

118th Congress: https://bit.ly/geography_118

Pennsylvania 2016 CD names are named by Shiro Kuriwaki and Lara Putnam.

Also see Cha, Kuriwaki, and Snyder, 2024, "Candidates in American General Elections", https://doi.org/10.7910/DVN/DGDRDT, Harvard Dataverse, for congressional candidate's (instead of President's) vote totals.

Details

cd_info_2008 is data on boundaries used in 2006, 2008, and 2010; cd_info_2012 is data on boundaries used in 2012 and 2014; cd_info_2016 uses 2016 boundaries; cd_info_2018 is data on 2018 boundaries; cd_info_2020 uses 2020 boundaries; cd_info_2022 is data on 2022 boundaries; cd_info_2024 uses 2024 boundaries.

District lines change before and after each decennial Census, e.g. 2010 vs. 2012 and 2020 vs. 2022. There is also change in district lines due to court interventions.

Between the 2022 and 2024 data, 4 districts changed their districts: AL, GA, LA, and NC. In these changes, AL-02, GA-06, and LA-06 became a majority minority district; the NC state supreme court plan in 2022 expired; and the NY court struck down the initial 2022 NY plan.

Between the 2018 and 2020 data, NC changed their districts

Between the 2016 and 2018 data, PA changed their districts

Between the 2012-14 and 2016 data, FL, NC, VA changed their districts

These can be seen by, for example, the following code: cd_info_2022 |> left_join(cd_info_2024, by = "cd") |> select(cd, matches("trump")) |> mutate(diff = abs(pct_trump - pct_trump20))

See also

Examples

head(cd_info_2018)
#> # A tibble: 6 × 11
#>    year cd    dailykos_name        largest_place pct_trump pct_romney pct_mccain
#>   <dbl> <chr> <chr>                <chr>             <dbl>      <dbl>      <dbl>
#> 1  2018 AK-01 All of Alaska        Anchorage         0.584      0.592      0.592
#> 2  2018 AL-01 Greater Mobile       Mobile            0.651      0.616      0.620
#> 3  2018 AL-02 Southeastern Alabama Montgomery        0.663      0.642      0.639
#> 4  2018 AL-03 Eastern Alabama      Auburn            0.669      0.630      0.630
#> 5  2018 AL-04 North-central Alaba… Gadsden           0.822      0.746      0.753
#> 6  2018 AL-05 Northern Alabama     Huntsville        0.674      0.638      0.642
#> # ℹ 4 more variables: presvotes_total_16 <dbl>, presvotes_total_20 <dbl>,
#> #   presvotes_DR_16 <dbl>, presvotes_DR_20 <dbl>
head(elec_NY)
#> # A tibble: 6 × 6
#>   cd    dailykos_name          largest_place pct_trump pct_romney pct_mccain
#>   <chr> <chr>                  <chr>             <dbl>      <dbl>      <dbl>
#> 1 NY-01 Eastern Suffolk County Coram             0.545      0.491       0.48
#> 2 NY-02 Southern Long Island   Brentwood         0.53       0.472       0.48
#> 3 NY-03 Northern Long Island   New York City     0.455      0.482       0.46
#> 4 NY-04 Long Island            Hempstead         0.438      0.428       0.44
#> 5 NY-05 Southeastern Queens    New York City     0.127      0.091       0.14
#> 6 NY-06 Northeastern Queens    New York City     0.321      0.31        0.36