Sample district level data

Some of the most consequential variables to include in MRP are at the district-level. We include one such data for congressional districts. All data is collected by Daily Kos.

cd_info_2018

Format

A dataframe with the 435 Congressional Districts, one row per cd.

cd_year

The year for the district line. A congressional district's actual geography can change year to year, and significantly so in different redistricting cylces. 2019 indicates the data about voteshare and placenames corresponds to district lines as of 2019.

cd

District code. The formatting corresponds to the CCES cumulative codeing of cd: a two-letter abbrevation for the state followed by a dash, and the district number padded with zeros to the left to be of length 2. At-large districts like Delwaware are given a "-01" for the district number.

pct_trump, pct_romney, pct_mccain

The voteshare of presidential candidates in that district for the given year. E.g. the pct_mccain data when cd_year == 2018 represents the percent of the vote by McCain in 2008 for that district under 2018 lines.

dailykos_name

The unique descriptive name for the district given by Daily Kos. See Source

largest_place

The largest place in the district. Multiple districts may have the largest place.

Source

The Daily Kos Elections naming guide to the nation's congressional districts. https://bit.ly/2XsFI5W

Daily Kos, "2008, 2012, & 2016 results for districts used in 2018." https://bit.ly/3bXtAPB

Examples

head(cd_info_2018)

#> # A tibble: 6 x 7
#>   cd_year cd    dailykos_name      largest_place pct_trump pct_romney pct_mccain
#>     <dbl> <chr> <chr>              <chr>             <dbl>      <dbl>      <dbl>
#> 1    2018 AK-01 All of Alaska      Anchorage         0.528      0.553      0.597
#> 2    2018 AL-01 Greater Mobile     Mobile            0.635      0.618      0.609
#> 3    2018 AL-02 Southeastern Alab… Montgomery        0.649      0.629      0.645
#> 4    2018 AL-03 Eastern Alabama    Auburn            0.653      0.623      0.626
#> 5    2018 AL-04 North-central Ala… Gadsden           0.804      0.748      0.733
#> 6    2018 AL-05 Northern Alabama   Huntsville        0.647      0.639      0.626