Download a specific CCES dataset from dataverse

You must download the development version of https://github.com/IQSS/dataverse-client-r and get your own dataverse API Key. The current dataverse package downloads the raw data, so this function writes the raw binary into a tempfile and loads it into a tibble with the appropriate file data type.

get_cces_dv(
  name = "cumulative",
  year_subset = 2006:2019,
  dataverse_paths = cces_dv_ids
)

Arguments

name
The name of the dataset as defined in data(cces_dv_ids).

year_subset
The year (or years, a vector) to subset too. If name is a year specific dataset, this argument is redundant, but if name == "cumulative", then the output will be the cumulative dataset subsetted to that year. This is useful when using the cumulative dataset for its harmonized variables.

dataverse_paths
A dataframe where one row represents metadata for one CCES dataset. Built-in data cces_dv_ids is used as a default and should not be changed.

Details

To get a user key, you need to have your own dataverse account. Go to your account online and look for "API Token". Generate a API Token. You can then save your token by following dataverse::get_user_key, or directly saving Sys.setenv("DATAVERSE_KEY" = "YOUR KEY HERE") Sys.setenv("DATAVERSE_SERVER" = "dataverse.harvard.edu")

Examples

if (FALSE) {
 ccc <- get_cces_dv("cumulative", year = 2018)
 }