Question metadata (sample)

Question metadata (sample)

questions_samp

Format

A tibble with 30 rows, each row corresponding to a variable (column) in a particular CCES dataset. It contains metadata about the question in columns.

q_ID

A unique identifier for the question. This is roughly a concatenation of cces_year and q_code, with some standardization.

q_label

A short descriptive label of the question, based on a reading of the actual question wording.

cces_data

The dataset the variable column comes from. Because the CCES is fielded each year, this is often a year, with the exception of "cumulative"

q_code

A string of the variable as it appears in the dataset.

response_type

A classification of the response option. Currently it consists of "yesno", "categorical" (discrete but not ordered), "ordinal" (discrete and ordered). All of these presume a discrete response.

Examples

 questions_samp
#> # A tibble: 30 × 5
#>    q_ID      q_label               cces_data q_code   response_type
#>    <chr>     <chr>                 <chr>     <chr>    <chr>        
#>  1 CC06_3075 Cut Capital Gains Tax 2006      v3075    yesno        
#>  2 CC06_3069 Immigration Reform    2006      v3069    yesno        
#>  3 CC07_34   SCHIP 2007            2007      CC34     yesno        
#>  4 CC08_316E SCHIP 2007            2008      CC316e   yesno        
#>  5 CC09_59F  PPACA                 2009      cc09_59f yesno        
#>  6 CC10_332D PPACA                 2010      CC332D   yesno        
#>  7 CC11_341D PPACA                 2011      CC341D   yesno        
#>  8 CC12_332G Repeal ACA 2012       2012      CC332G   yesno        
#>  9 CC12_332I PPACA                 2012      CC332I   yesno        
#> 10 CC12_332D Tax Relief 2012       2012      CC332D   yesno        
#> # … with 20 more rows