CCES-ACS variable recode keys

CCES-ACS variable recode keys

race_key

gender_key

educ_key

educ_key2

age5_key

age10_key

Format

A tibble with one row per recoding value

race, educ, gender

An labelled integer of class haven::labelled. Most compact form of both sources and the values both will get recoded to in MRP.

race_chr, gender_chr, educ_chr

Labels for the first column, in characters

race_cces_chr, edu_cces_chr

Corresponding labelled in the CCES cumulative common content

race_acs

Corresponding character in the ACS data via the tidycensus package

An object of class tbl_df (inherits from tbl, data.frame) with 2 rows and 2 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 7 rows and 3 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 7 rows and 3 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 5 rows and 2 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 10 rows and 3 columns.

Details

These value-value tables are useful for recoding the values of from one dataset (CCES) so that they can be merged immediately with another (ACS). These get used internally in ccc_std_demographics, but they are available as built in datasets

Examples

 library(ccesMRPprep)
 educ_key

#> # A tibble: 7 x 3
#>   educ_chr                                  educ_cces_chr                   educ
#>   <chr>                                     <chr>                      <dbl+lbl>
#> 1 Less than 9th grade                       No HS                1 [HS or Less] 
#> 2 9th to 12th grade no diploma              No HS                1 [HS or Less] 
#> 3 High school graduate (includes equivalen… High School Graduate 1 [HS or Less] 
#> 4 Some college no degree                    Some College         2 [Some Colleg…
#> 5 Associate's degree                        2-Year               2 [Some Colleg…
#> 6 Bachelor's degree                         4-Year               3 [4-Year]     
#> 7 Graduate or professional degree           Post-Grad            4 [Post-Grad]  

 race_key

#> # A tibble: 9 x 5
#>   race_5 race_cces_chr   race_acs                           race_cces       race
#>    <int> <chr>           <chr>                              <int+lbl>  <int+lbl>
#> 1      1 White           WHITE ALONE, NOT HISPANIC OR LA…  1 [White]  1 [White] 
#> 2      2 Black           BLACK OR AFRICAN AMERICAN ALONE   2 [Black]  2 [Black] 
#> 3      3 Hispanic        HISPANIC OR LATINO                3 [Hispan… 3 [Hispan…
#> 4      4 Asian           ASIAN ALONE                       4 [Asian]  4 [Asian] 
#> 5      4 Asian           NATIVE HAWAIIAN AND OTHER PACIF…  4 [Asian]  4 [Asian] 
#> 6      5 Native American AMERICAN INDIAN AND ALASKA NATI… NA          5 [All Ot…
#> 7      5 Mixed           TWO OR MORE RACES                NA          5 [All Ot…
#> 8      5 Other           SOME OTHER RACE ALONE            NA          5 [All Ot…
#> 9      5 Middle Eastern  NA                               NA          5 [All Ot…