Sample 2018 Common Content

1000 rows from th 2018 CCES Common Content. Contains all columns, retaining the haven_labelled class. This is doi:10.7910/DVN/ZSBZ7K/H5IDTA originally read in using read_dta via get_cces_dataverse(). To search for variable by its content, questionr::lookfor (or rcces::vartab) is a useful option (see example).

cc18_samp

Format

An object of class tbl_df (inherits from tbl, data.frame) with 1000 rows and 526 columns.

Source

"Brian Schaffner; Stephen Ansolabehere; Sam Luks, 2019, "CCES Common Content, 2018", doi:https://doi.org/10.7910, Harvard Dataverse V6.

Details

See the 2018 Codebook in the DOI below for question wording of each column. To use the harmonized and easier-to-use versions of common variables like geography, partisan ID, demographics and vote choice, use the cumulative common content, of which 2018 is a subset. A sample of the cumulative is contained in this package as ?ccc_samp.

Examples

# use questionr::lookfor to search the label and labels
questionr::lookfor(cc18_samp, "Trump")

#>  pos variable            label                 col_type values               
#>  91  CC18_308a           Job approval -- Pres~ dbl+lbl  [1] Strongly approve 
#>                                                         [2] Somewhat approve 
#>                                                         [3] Somewhat disappr~
#>                                                         [4] Strongly disappr~
#>                                                         [5] Not sure         
#>                                                         [8] skipped          
#>                                                         [9] not asked        
#>  149 CC18_334C           Ideological Placemen~ dbl+lbl  [1] Very Liberal     
#>                                                         [2] Liberal          
#>                                                         [3] Somewhat Liberal 
#>                                                         [4] Middle of the Ro~
#>                                                         [5] Somewhat Conserv~
#>                                                         [6] Conservative     
#>                                                         [7] Very Conservative
#>                                                         [8] Not sure         
#>                                                         [98] skipped         
#>                                                         [99] not asked       
#>  162 CC18_335            Trump Russia collusi~ dbl+lbl  [1] Yes              
#>                                                         [2] No               
#>                                                         [3] Not sure         
#>                                                         [8] skipped          
#>                                                         [9] not asked        
#>  430 CC18_app_dtrmp_post President Trump Job ~ dbl+lbl  [1] Strongly approve 
#>                                                         [2] Somewhat approve 
#>                                                         [3] Somewhat disappr~
#>                                                         [4] Strongly disappr~
#>                                                         [5] Not sure         
#>                                                         [8] skipped          
#>                                                         [9] not asked        

# all data
cc18_samp

#> # A tibble: 1,000 x 526
#>     year   case_id commonweight commonpostweight vvweight vvweight_post tookpost
#>    <dbl>     <dbl>        <dbl>            <dbl>    <dbl>         <dbl>    <dbl>
#>  1  2018 415395741        1.47             1.48    NA            NA            2
#>  2  2018 414164923        1.33             1.22     1.51          1.29         2
#>  3  2018 412379892        0.641            0.559    0.602         0.594        2
#>  4  2018 414203529        1.35             1.19     1.18          1.17         2
#>  5  2018 412148048        0.870            0.762    0.842         0.742        2
#>  6  2018 412329835        0.949            0.789    1.07          0.983        2
#>  7  2018 417352072        0.688            0.585    0.717         0.716        2
#>  8  2018 414614677        0.814            0.616   NA            NA            2
#>  9  2018 416797006        1.12             0.926    0.991         0.929        2
#> 10  2018 412962561        1.05             0.877    0.346         0.372        2
#> # … with 990 more rows, and 519 more variables: CCEStake <dbl>, birthyr <dbl>,
#> #   gender <dbl>, educ <dbl>, race <dbl>, race_other <chr>, hispanic <dbl>,
#> #   marstat <dbl>, multrace_1 <dbl>, multrace_2 <dbl>, multrace_3 <dbl>,
#> #   multrace_4 <dbl>, multrace_5 <dbl>, multrace_8 <dbl>, multrace_97 <dbl>,
#> #   multrace_98 <dbl>, multrace_99 <dbl>, CC18_354a_1 <dbl>, CC18_354a_2 <dbl>,
#> #   CC18_354a_3 <dbl>, CC18_354a_4 <dbl>, CC18_354a_5 <dbl>, CC18_354a_6 <dbl>,
#> #   CC18_354a_7 <dbl>, CC18_354a_8 <dbl>, CC18_354a_9 <dbl>,
#> #   CC18_354a_10 <dbl>, CC18_354a_11 <dbl>, CC18_354a_12 <dbl>,
#> #   CC18_354a_t <chr>, CC18_354b_1 <dbl>, CC18_354b_2 <dbl>, CC18_354b_3 <dbl>,
#> #   CC18_354b_4 <dbl>, CC18_354b_5 <dbl>, CC18_354b_6 <dbl>, CC18_354b_7 <dbl>,
#> #   CC18_354b_8 <dbl>, CC18_354b_9 <dbl>, CC18_354b_10 <dbl>,
#> #   CC18_354b_11 <dbl>, CC18_354b_12 <dbl>, CC18_354b_13 <dbl>,
#> #   CC18_354b_14 <dbl>, CC18_354b_15 <dbl>, CC17_3534_t <chr>, comptype <dbl>,
#> #   add_confirm <dbl>, inputzip <dbl>, votereg <dbl>, votereg_f <dbl>,
#> #   regzip <dbl>, OScode <dbl>, CompRating <dbl>, region <dbl>,
#> #   sample_group <dbl>, ccesmodule <dbl>, CC18_300_1 <dbl>, CC18_300_2 <dbl>,
#> #   CC18_300_3 <dbl>, CC18_300_4 <dbl>, CC18_300_5 <dbl>, CC18_300_6 <dbl>,
#> #   CC18_300b <dbl>, CC18_300c <dbl>, CC18_300d_1 <dbl>, CC18_300d_2 <dbl>,
#> #   CC18_300d_3 <dbl>, CC18_300d_4 <dbl>, CC18_300d_5 <dbl>, CC18_301 <dbl>,
#> #   CC18_302 <dbl>, CC18_303_1 <dbl>, CC18_303_2 <dbl>, CC18_303_3 <dbl>,
#> #   CC18_303_4 <dbl>, CC18_303_5 <dbl>, CC18_303_6 <dbl>, CC18_303_7 <dbl>,
#> #   CC18_303_8 <dbl>, CC18_303_9 <dbl>, CC18_303_10 <dbl>, CC18_303_11 <dbl>,
#> #   CC18_308a <dbl>, CC18_308b <dbl>, CC18_308c <dbl>, CC18_308d <dbl>,
#> #   CC18_308e <dbl>, CC18_309a <dbl>, CC18_309b <dbl>, CC18_309c <dbl>,
#> #   CC18_309d <dbl>, CC18_310a <dbl>, CC18_310b <dbl>, CC18_310c <dbl>,
#> #   CC18_310d <dbl>, CC18_311a <dbl>, CC18_311b <dbl>, CC18_311c <dbl>,
#> #   CC18_316 <dbl>, …