Sample 2018 Common Content

1000 rows from th 2018 CCES Common Content. Contains all columns, retaining the haven_labelled class. This is doi:10.7910/DVN/ZSBZ7K/H5IDTA originally read in using read_dta via get_cces_dataverse(). To search for variable by its content, questionr::lookfor (or rcces::vartab) is a useful option (see example).

cc18_samp

Format

An object of class tbl_df (inherits from tbl, data.frame) with 1000 rows and 526 columns.

Source

"Brian Schaffner; Stephen Ansolabehere; Sam Luks, 2019, "CCES Common Content, 2018", doi:https://doi.org/10.7910, Harvard Dataverse V6.

Details

See the 2018 Codebook in the DOI below for question wording of each column. To use the harmonized and easier-to-use versions of common variables like geography, partisan ID, demographics and vote choice, use the cumulative common content, of which 2018 is a subset. A sample of the cumulative is contained in this package as ?ccc_samp.

Examples

# use questionr::lookfor to search the label and labels
questionr::lookfor(cc18_samp, "Trump")
#>  pos variable            label                 col_type values               
#>  91  CC18_308a           Job approval -- Pres~ dbl+lbl  [1] Strongly approve 
#>                                                         [2] Somewhat approve 
#>                                                         [3] Somewhat disappr~
#>                                                         [4] Strongly disappr~
#>                                                         [5] Not sure         
#>                                                         [8] skipped          
#>                                                         [9] not asked        
#>  108 CC18_317            President 2016        dbl+lbl  [1] Donald Trump     
#>                                                         [2] Hillary Clinton  
#>                                                         [3] Someone else     
#>                                                         [4] I did not cast a~
#>                                                         [5] I don't recall   
#>                                                         [8] skipped          
#>                                                         [9] not asked        
#>  149 CC18_334C           Ideological Placemen~ dbl+lbl  [1] Very Liberal     
#>                                                         [2] Liberal          
#>                                                         [3] Somewhat Liberal 
#>                                                         [4] Middle of the Ro~
#>                                                         [5] Somewhat Conserv~
#>                                                         [6] Conservative     
#>                                                         [7] Very Conservative
#>                                                         [8] Not sure         
#>                                                         [98] skipped         
#>                                                         [99] not asked       
#>  162 CC18_335            Trump Russia collusi~ dbl+lbl  [1] Yes              
#>                                                         [2] No               
#>                                                         [3] Not sure         
#>                                                         [8] skipped          
#>                                                         [9] not asked        
#>  430 CC18_app_dtrmp_post President Trump Job ~ dbl+lbl  [1] Strongly approve 
#>                                                         [2] Somewhat approve 
#>                                                         [3] Somewhat disappr~
#>                                                         [4] Strongly disappr~
#>                                                         [5] Not sure         
#>                                                         [8] skipped          
#>                                                         [9] not asked        

# all data
cc18_samp
#> # A tibble: 1,000 × 526
#>     year   case_id commonweight commonpostweight vvweight vvweight_post tookpost
#>    <dbl>     <dbl>        <dbl>            <dbl>    <dbl>         <dbl> <dbl+lb>
#>  1  2018 415395741        1.47             1.48    NA            NA      2 [Yes]
#>  2  2018 414164923        1.33             1.22     1.51          1.29   2 [Yes]
#>  3  2018 412379892        0.641            0.559    0.602         0.594  2 [Yes]
#>  4  2018 414203529        1.35             1.19     1.18          1.17   2 [Yes]
#>  5  2018 412148048        0.870            0.762    0.842         0.742  2 [Yes]
#>  6  2018 412329835        0.949            0.789    1.07          0.983  2 [Yes]
#>  7  2018 417352072        0.688            0.585    0.717         0.716  2 [Yes]
#>  8  2018 414614677        0.814            0.616   NA            NA      2 [Yes]
#>  9  2018 416797006        1.12             0.926    0.991         0.929  2 [Yes]
#> 10  2018 412962561        1.05             0.877    0.346         0.372  2 [Yes]
#> # … with 990 more rows, and 519 more variables: CCEStake <dbl+lbl>,
#> #   birthyr <dbl>, gender <dbl+lbl>, educ <dbl+lbl>, race <dbl+lbl>,
#> #   race_other <chr>, hispanic <dbl+lbl>, marstat <dbl+lbl>,
#> #   multrace_1 <dbl+lbl>, multrace_2 <dbl+lbl>, multrace_3 <dbl+lbl>,
#> #   multrace_4 <dbl+lbl>, multrace_5 <dbl+lbl>, multrace_8 <dbl+lbl>,
#> #   multrace_97 <dbl+lbl>, multrace_98 <dbl+lbl>, multrace_99 <dbl+lbl>,
#> #   CC18_354a_1 <dbl+lbl>, CC18_354a_2 <dbl+lbl>, CC18_354a_3 <dbl+lbl>, …