Sample cumulative CCES data

The cumulative CCES stacks CCES common content for all years and harmonizes the variables, which makes it ideal for using it for MRP. This is a sample for illustration; see get_cces_dv to get the full data.

ccc_samp

Format

A data frame with 1000 observations. See the CCES cumulative codebook for more explanation of the variables

year

Year of the common content

case_id

Respondent identifier (unique within each year)

state

State (in the form of state.name)

st

State Abbreviation (in the form of state.abb)

cd

Congressional district at the time of the survey. See ccc_std_demographics for how it is and should be standardized.

zipcode

Zipcode (See codebook)

county_fips

County FIPS code (See codebook)

gender

Gender (equivalent to sex in ACS for the purposes of this package)

age

Age, in integers

race

Race and ethnicity

hispanic

Hispanic identifier

educ

Education level

faminc

Family Income

marstat

Martial status

newsint

Frequency of following the news

vv_turnout_gvm

Validated turnout in general election

voted_pres_16

Self-reported vote choice for 2016

economy_retro

Opinion on retrospective economy

Source

Kuriwaki, Shiro, 2018, "Cumulative CCES Common Content (2006-2018)", https://doi.org/10.7910/DVN/II2DB6, Harvard Dataverse

Details

This is encoded as a RDS file with some variables stored in the Stata-based integer + labelled class instead of as factors. See the CCES cumulative guide for the difference between the two and how to go from one to another.

Examples

library(ccesMRPprep)
ccc_samp

#> # A tibble: 1,000 x 18
#>     year case_id state st    cd    zipcode county_fips  gender   age    race
#>    <dbl> <chr>   <chr> <chr> <chr> <chr>   <chr>       <dbl+l> <dbl> <dbl+l>
#>  1  2006 1005058 Mich… MI    MI-04 48603   26145       2 [Fem…    36 1 [Whi…
#>  2  2006 1006614 Texas TX    TX-18 77040   48201       1 [Mal…    40 3 [His…
#>  3  2006 1009338 Cali… CA    CA-48 92656   06059       1 [Mal…    32 1 [Whi…
#>  4  2006 1088898 Flor… FL    FL-13 34224   12015       2 [Fem…    52 1 [Whi…
#>  5  2006 1090564 Penn… PA    PA-19 17011   42041       2 [Fem…    25 1 [Whi…
#>  6  2006 1093132 Sout… SC    SC-02 29073   45063       2 [Fem…    48 1 [Whi…
#>  7  2006 1093573 Utah  UT    UT-03 84118   49035       2 [Fem…    74 1 [Whi…
#>  8  2006 1105620 Hawa… HI    HI-01 96701   15003       1 [Mal…    37 4 [Asi…
#>  9  2006 1116569 Texas TX    TX-21 78610   48209       2 [Fem…    20 3 [His…
#> 10  2006 1117377 Ohio  OH    OH-07 45502   39023       2 [Fem…    49 1 [Whi…
#> # … with 990 more rows, and 8 more variables: hispanic <dbl+lbl>,
#> #   educ <dbl+lbl>, faminc <dbl+lbl>, marstat <dbl+lbl>, newsint <dbl+lbl>,
#> #   vv_turnout_gvm <dbl+lbl>, voted_pres_16 <dbl+lbl>, economy_retro <dbl+lbl>