Recode CCES variables so that they merge to ACS variables

Recode CCES variables so that they merge to ACS variables

ccc_std_demographics(tbl, only_demog = FALSE)

Arguments

tbl
A subset of the cumulative common content. Must include variables age, race, educ, and gender.

only_demog
Drop variables besides demographics? Defaults to FALSE

Value

The output is of the same dimensions as the input (unless only_demog = TRUE) but with the following exceptions. age_bin is coded to match up with the ACS bins and the recoding occurs in a separate function, ccc_bin_age. educ is recoded (coarsened and relabelled) to match up with the ACS. (the original version is left as educ_cces_chr), and the same goes for race. These recodings are governed by the key-value pairs educ_key and race_key.

Examples

if (FALSE) {
 library(dataverse)

 # Load Cumulative dataset (need your own Token)
 cumulative_rds <- get_file(file = "cumulative_2006_2018.Rds",
                            dataset = "doi:10.7910/DVN/II2DB6")
 tmp <- tempfile(fileext = ".Rds")
 writeBin(cumulative_rds, tmp)
 cumulative_rds <- readr::read_rds(tmp)

 cumulative_std <- ccc_std_demographics(cumulative_rds)

 }