Recode CCES variables so that they merge to ACS variables

Recode CCES variables so that they merge to ACS variables

ccc_std_demographics(
  tbl,
  only_demog = FALSE,
  age_key = deframe(ccesMRPprep::age5_key)
)

Arguments

tbl
A subset of the cumulative common content. Must include variables age, race, educ, gender, st, state, and cd. See ccc_samp for an example.

only_demog
Drop variables besides demographics? Defaults to FALSE

age_key
The vector key to use to bin age. Can be deframe(age5_key) or deframe(age10_key)

Value

The output is of the same dimensions as the input (unless only_demog = TRUE) but with the following exceptions:

age is coded to match up with the ACS bins and the recoding occurs in a separate function, ccc_bin_age. The unbinned age is left instead to age_orig.

educ is recoded (coarsened and relabelled) to match up with the ACS. (the original version is left as educ_cces_chr). Recoding is governed by the key-value pairs educ_key

the same goes for race. These recodings are governed by the key-value pair race_key.

cd is standardized so that at large districs are given "01" and single-digit districts are padded with 0s. e.g. "WY-01" and "CA-02".

Examples

if (FALSE) {
 library(dataverse)

 # Load Cumulative dataset (need your own Token)
 cumulative_rds <- get_cces_dv("cumulative")

 cumulative_std <- ccc_std_demographics(cumulative_rds)

 }