Identifiers for datasets in CCES Dataverse

Identifiers for datasets in CCES Dataverse

cces_dv_ids

Format

A tibble with one row per dataverse dataset. The doi, filename, and server can uniquely define a file in ?dataverse::get_file().

cces_name

The unique name/shorthand for this dataset

year

Specific year the data surveys, if there is one unique year

doi

The DOI for the dataverse dataset (instead of the specific file). The URL is therefore https://doi.org/ followed by the DOI

filename

The name of the data file, as it appears in the dataverse.

server

The particular dataverse instance.

caseid_var

The variable in that dataset that specifies the unique case (respondent/person) identifier. Use a string.

Examples

 cces_dv_ids

#> # A tibble: 16 x 6
#>    cces_name   year doi          filename               caseid_var server       
#>    <chr>      <dbl> <chr>        <chr>                  <chr>      <chr>        
#>  1 cumulative    NA 10.7910/DVN… cumulative_2006-2020.… case_id    dataverse.ha…
#>  2 2006        2006 10.7910/DVN… cces_2006_common.tab   v1000      dataverse.ha…
#>  3 2007        2007 10.7910/DVN… CCES07_OUTPUT.sav      caseid     dataverse.ha…
#>  4 2008        2008 10.7910/DVN… cces_2008_common.tab   V100       dataverse.ha…
#>  5 2009        2009 10.7910/DVN… cces09_cmn_output_2.d… v100       dataverse.ha…
#>  6 2010        2010 10.7910/DVN… cces_2010_common_vali… V100       dataverse.ha…
#>  7 2011        2011 10.7910/DVN… CCES11_Common_OUTPUT.… V100       dataverse.ha…
#>  8 2012        2012 10.7910/DVN… CCES12_Common_VV.tab   V101       dataverse.ha…
#>  9 2013        2013 10.7910/DVN… Common Content Data.t… caseid     dataverse.ha…
#> 10 2014        2014 10.7910/DVN… CCES14_Common_Content… V101       dataverse.ha…
#> 11 2015        2015 10.7910/DVN… CCES15_Common_OUTPUT_… V101       dataverse.ha…
#> 12 2016        2016 10.7910/DVN… CCES16_Common_OUTPUT_… V101       dataverse.ha…
#> 13 2017        2017 10.7910/DVN… Common Content Data.t… V101       dataverse.ha…
#> 14 2018        2018 10.7910/DVN… cces18_common_vv.dta   caseid     dataverse.ha…
#> 15 2019        2019 10.7910/DVN… CCES19_Common_OUTPUT.… caseid     dataverse.ha…
#> 16 2020        2020 10.7910/DVN… CCES20_Common_OUTPUT.… caseid     dataverse.ha…