Congressional District level information by Daily Kos

Some of the most consequential variables to include in MRP are at the district-level. We include one such data for congressional districts. All data is collected by Daily Kos. cd_info_2018 is data on 2018 boundaries, cd_info_2016 uses 2016 boundaries and cd_info_2020 uses 2020 (but with place descriptions currently at 2016).

cd_info_2018

cd_info_2016

cd_info_2020

elec_NY

Format

Each cd_info_20** is a dataframe with the 435 Congressional Districts, one row per cd.

year

The year for the district line. A congressional district's actual geography can change year to year, and significantly so in different redistricting cycles. Lines try to get the contemporaneous district map, so that cd_info_2016 uses 2016 maps and cd_info_2020 uses 2020 maps. However, this work relies on the hard work of assembling precinct results by Daily Kos.

cd

District code. The formatting corresponds to the CCES cumulative coding of cd: a two-letter abbreviation for the state followed by a dash, and the district number padded with zeros to the left to be of length 2. At-large districts like Delaware are given a "-01" for the district number.

presvotes_total

In presidential years, the total number of votes cast for the office of President that year. Taken from Daily Kos estimates from precinct results.

pct_trump, pct_romney, pct_mccain

The two-party voteshare of Republican presidential candidates in that district for the given year. E.g. the pct_mccain data when cd_year == 2018 represents the percent of the vote by McCain in 2008 for that district under 2018 lines. The Trump value is for 2016 for cd_info_2018 and ``cd_info_2020but not for 2020 where we use Trump's 2020 vote against Biden and denote aspct_trump16` the 2016 result.

dailykos_name

The unique descriptive name for the district code in 2018 given by Daily Kos. Some edits are made for changing district. See Source for full citation.

largest_place

The largest place in the district code in 2018 given by Daily Kos. Multiple districts may have the largest place.

An object of class tbl_df (inherits from tbl, data.frame) with 435 rows and 7 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 435 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 27 rows and 6 columns.

Source

The Daily Kos Elections naming guide to the nation's congressional districts. https://bit.ly/2XsFI5W

Daily Kos, "2008, 2012, & 2016 results for districts used in 2018." https://bit.ly/3DRhPcj DailyDaily Kos Elections 2012, 2016 & 2020 presidential election results for congressional districts in 2020" https://bit.ly/3bXtAPB

Pennsylvania 2016 CD names are named by Shiro Kuriwaki and Lara Putnam.

Also see Cha, Jeremiah; Kuriwaki, Shiro; Snyder, James M. Jr., 2021, "Candidates in American General Elections", https://doi.org/10.7910/DVN/DGDRDT, Harvard Dataverse.

Examples

head(cd_info_2018)
#> # A tibble: 6 × 7
#>    year cd    dailykos_name        largest_place pct_trump pct_romney pct_mccain
#>   <dbl> <chr> <chr>                <chr>             <dbl>      <dbl>      <dbl>
#> 1  2018 AK-01 All of Alaska        Anchorage         0.584      0.592      0.592
#> 2  2018 AL-01 Greater Mobile       Mobile            0.651      0.616      0.620
#> 3  2018 AL-02 Southeastern Alabama Montgomery        0.663      0.642      0.639
#> 4  2018 AL-03 Eastern Alabama      Auburn            0.669      0.630      0.630
#> 5  2018 AL-04 North-central Alaba… Gadsden           0.822      0.746      0.753
#> 6  2018 AL-05 Northern Alabama     Huntsville        0.674      0.638      0.642
head(elec_NY)
#> # A tibble: 6 × 6
#>   cd    dailykos_name          largest_place pct_trump pct_romney pct_mccain
#>   <chr> <chr>                  <chr>             <dbl>      <dbl>      <dbl>
#> 1 NY-01 Eastern Suffolk County Coram             0.545      0.491       0.48
#> 2 NY-02 Southern Long Island   Brentwood         0.53       0.472       0.48
#> 3 NY-03 Northern Long Island   New York City     0.455      0.482       0.46
#> 4 NY-04 Long Island            Hempstead         0.438      0.428       0.44
#> 5 NY-05 Southeastern Queens    New York City     0.127      0.091       0.14
#> 6 NY-06 Northeastern Queens    New York City     0.321      0.31        0.36