CCES-ACS variable key value pairs for recoding

These value-value tables are useful for recoding the values of from one dataset (CCES) so that they can be merged immediately with another (ACS). These get used internally in ccc_std_demographics, but they are available as built in datasets.

race_key

gender_key

educ_key

age5_key

age10_key

states_key

Format

All keys are tibbles with one row per recoding value.

race_key

race

An labelled integer of class haven::labelled. Most compact form of both sources and the values both will get recoded to in MRP.

race_cces

Labelled versions of the CCES race codings. These are of the same class as the CCES cumulative file.

race_cces_chr

Labels for the first column, in characters

race_acs

Corresponding character in the ACS data via the tidycensus package

race_5

A numeric value underlying the race label.

gender_key:

gender

An labelled integer of class haven::labelled. Target variable

gender_chr

Character to recode from. CCES and ACS use the same values.

educ_key

For mapping ACS data values for education e.g. in get_acs_cces:

educ_cces_chr

Character to recode from, in CCES

educ_chr

Character to recode from, in ACS.

educ

An labelled integer of class haven::labelled. Target variable

age5_key

Age bins, 5-ways, used in acscodes_age_sex_educ. Use ccc_bin_age to recode CCES variable

age

An labelled integer of class haven::labelled. Target variable.

age_chr

Character to recode from, in ACS

age10_key: Age bins, 10-ways, used in acscodes_age_sex_race:

age

An labelled integer of class haven::haven_labelled. Target variable.

age_chr

Character to recode from, in ACS

states_key: State codes and regions:

st

State two-letter abbreviation state.abb

state

State full name via state.name

st_fips

Integer, state FIPS code

region

Census region (Northeast, Midwest, South, West)

division

Census division (New England, Middle Atlantic, South Atlantic, East South Central, West South Central, East North Central, "West North Central, Mountain, Pacific)

Details

These tibbles themselves are not key-values pair in a strict sense because the dataframe tries to have two recodes CCES to common and ACS to common and so for a given recode, rows are not distinct. To avoid duplicating rows inadvertently, use the dplyr::distinct to reduce the key to two columns with unique rows.

Examples

 library(ccesMRPprep)
 race_key

#> # A tibble: 9 x 5
#>   race_5 race_cces_chr   race_acs                           race_cces       race
#>    <int> <chr>           <chr>                              <int+lbl>  <int+lbl>
#> 1      1 White           WHITE ALONE, NOT HISPANIC OR LA…  1 [White]  1 [White] 
#> 2      2 Black           BLACK OR AFRICAN AMERICAN ALONE   2 [Black]  2 [Black] 
#> 3      3 Hispanic        HISPANIC OR LATINO                3 [Hispan… 3 [Hispan…
#> 4      4 Asian           ASIAN ALONE                       4 [Asian]  4 [Asian] 
#> 5      4 Asian           NATIVE HAWAIIAN AND OTHER PACIF…  4 [Asian]  4 [Asian] 
#> 6      5 Native American AMERICAN INDIAN AND ALASKA NATI… NA          5 [All Ot…
#> 7      5 Mixed           TWO OR MORE RACES                NA          5 [All Ot…
#> 8      5 Other           SOME OTHER RACE ALONE            NA          5 [All Ot…
#> 9      5 Middle Eastern  NA                               NA          5 [All Ot…

 educ_key

#> # A tibble: 10 x 3
#>    educ_chr                                  educ_cces_chr                  educ
#>    <chr>                                     <chr>                     <dbl+lbl>
#>  1 Less than 9th grade                       No HS               1 [HS or Less] 
#>  2 9th to 12th grade no diploma              No HS               1 [HS or Less] 
#>  3 9th to 12th grade, no diploma             No HS               1 [HS or Less] 
#>  4 High school graduate (includes equivalen… High School Gradua… 1 [HS or Less] 
#>  5 High school graduate, GED, or alternative High School Gradua… 1 [HS or Less] 
#>  6 Some college no degree                    Some College        2 [Some Colleg…
#>  7 Some college, no degree                   Some College        2 [Some Colleg…
#>  8 Associate's degree                        2-Year              2 [Some Colleg…
#>  9 Bachelor's degree                         4-Year              3 [4-Year]     
#> 10 Graduate or professional degree           Post-Grad           4 [Post-Grad]  

 gender_key

#> # A tibble: 2 x 2
#>   gender_chr     gender
#>   <chr>       <int+lbl>
#> 1 Male       1 [Male]  
#> 2 Female     2 [Female]

 age5_key

#> # A tibble: 5 x 2
#>   age_chr                             age
#>   <chr>                         <int+lbl>
#> 1 18 to 24 years    1 [18 to 24 years]   
#> 2 25 to 34 years    2 [25 to 34 years]   
#> 3 35 to 44 years    3 [35 to 44 years]   
#> 4 45 to 64 years    4 [45 to 64 years]   
#> 5 65 years and over 5 [65 years and over]

 age10_key

#> # A tibble: 10 x 3
#>    age_chr           age_num                   age
#>    <chr>               <int>             <int+lbl>
#>  1 18 and 19 years         1 1 [18 to 24 years]   
#>  2 20 to 24 years          1 1 [18 to 24 years]   
#>  3 25 to 29 years          2 2 [25 to 34 years]   
#>  4 30 to 34 years          2 2 [25 to 34 years]   
#>  5 35 to 44 years          3 3 [35 to 44 years]   
#>  6 45 to 54 years          4 4 [45 to 64 years]   
#>  7 55 to 64 years          4 4 [45 to 64 years]   
#>  8 65 to 74 years          5 5 [65 years and over]
#>  9 75 to 84 years          5 5 [65 years and over]
#> 10 85 years and over       5 5 [65 years and over]

 states_key

#> # A tibble: 50 x 5
#>    st    state       st_fips region    division          
#>    <chr> <chr>         <int> <fct>     <fct>             
#>  1 AL    Alabama           1 South     East South Central
#>  2 AK    Alaska            2 West      Pacific           
#>  3 AZ    Arizona           4 West      Mountain          
#>  4 AR    Arkansas          5 South     West South Central
#>  5 CA    California        6 West      Pacific           
#>  6 CO    Colorado          8 West      Mountain          
#>  7 CT    Connecticut       9 Northeast New England       
#>  8 DE    Delaware         10 South     South Atlantic    
#>  9 FL    Florida          12 South     South Atlantic    
#> 10 GA    Georgia          13 South     South Atlantic    
#> # … with 40 more rows